requirements.txt
setuptools
pyinstaller
auto-py-to-exe
urllib3
PyQt5
PyQtWebEngine
colorama
psutil
keyboard
requests
dnspython
deep-translator
phonenumbers
bcrypt
beautifulsoup4
pypiwin32
whois
cryptography
screeninfo
GPUtil
pycryptodome
discord
discord.py
Pillow
browser-cookie3
opencv-python
pyautogui
beautifulsoup4
lxml